+++ bmc-dataflow/src/main/java/com/oracle/bmc/dataflow/model/Run.java
@@ -398,7 +398,7 @@ public class Run {
     String className;
 
     /**
-     * The OCID of the compartment that contains this application.
+     * The OCID of a compartment.
      *
      **/
     @com.fasterxml.jackson.annotation.JsonProperty("compartmentId")