+++ bmc-keymanagement/src/main/java/com/oracle/bmc/keymanagement/KmsVaultClient.java
@@ -540,6 +540,34 @@ public class KmsVaultClient implements KmsVault {
     }
 
     @Override
+    public GetVaultUsageResponse getVaultUsage(GetVaultUsageRequest request) {
+        LOG.trace("Called getVaultUsage");
+        final GetVaultUsageRequest interceptedRequest =
+                GetVaultUsageConverter.interceptRequest(request);
+        com.oracle.bmc.http.internal.WrappedInvocationBuilder ib =
+                GetVaultUsageConverter.fromRequest(client, interceptedRequest);
+        com.google.common.base.Function<javax.ws.rs.core.Response, GetVaultUsageResponse>
+                transformer = GetVaultUsageConverter.fromResponse();
+
+        final com.oracle.bmc.retrier.BmcGenericRetrier retrier =
+                com.oracle.bmc.retrier.Retriers.createPreferredRetrier(
+                        interceptedRequest.getRetryConfiguration(), retryConfiguration);
+        return retrier.execute(
+                interceptedRequest,
+                retryRequest -> {
+                    final com.oracle.bmc.retrier.TokenRefreshRetrier tokenRefreshRetrier =
+                            new com.oracle.bmc.retrier.TokenRefreshRetrier(
+                                    authenticationDetailsProvider);
+                    return tokenRefreshRetrier.execute(
+                            retryRequest,
+                            retriedRequest -> {
+                                javax.ws.rs.core.Response response = client.get(ib, retriedRequest);
+                                return transformer.apply(response);
+                            });
+                });
+    }
+
+    @Override
     public ListVaultsResponse listVaults(ListVaultsRequest request) {
         LOG.trace("Called listVaults");
         final ListVaultsRequest interceptedRequest = ListVaultsConverter.interceptRequest(request);