+++ bmc-streaming/src/main/java/com/oracle/bmc/streaming/model/PrivateEndpointDetails.java
@@ -1,0 +1,114 @@
+/**
+ * Copyright (c) 2016, 2020, Oracle and/or its affiliates. All rights reserved.
+ */
+package com.oracle.bmc.streaming.model;
+
+/**
+ * Optional parameters if a private stream pool is requested.
+ * <br/>
+ * Note: Objects should always be created or deserialized using the {@link Builder}. This model distinguishes fields
+ * that are {@code null} because they are unset from fields that are explicitly set to {@code null}. This is done in
+ * the setter methods of the {@link Builder}, which maintain a set of all explicitly set fields called
+ * {@link #__explicitlySet__}. The {@link #hashCode()} and {@link #equals(Object)} methods are implemented to take
+ * {@link #__explicitlySet__} into account. The constructor, on the other hand, does not set {@link #__explicitlySet__}
+ * (since the constructor cannot distinguish explicit {@code null} from unset {@code null}).
+ **/
+@javax.annotation.Generated(value = "OracleSDKGenerator", comments = "API Version: 20180418")
+@lombok.AllArgsConstructor(onConstructor = @__({@Deprecated}))
+@lombok.Value
+@com.fasterxml.jackson.databind.annotation.JsonDeserialize(
+    builder = PrivateEndpointDetails.Builder.class
+)
+@com.fasterxml.jackson.annotation.JsonFilter(com.oracle.bmc.http.internal.ExplicitlySetFilter.NAME)
+public class PrivateEndpointDetails {
+    @com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder(withPrefix = "")
+    @lombok.experimental.Accessors(fluent = true)
+    public static class Builder {
+        @com.fasterxml.jackson.annotation.JsonProperty("subnetId")
+        private String subnetId;
+
+        public Builder subnetId(String subnetId) {
+            this.subnetId = subnetId;
+            this.__explicitlySet__.add("subnetId");
+            return this;
+        }
+
+        @com.fasterxml.jackson.annotation.JsonProperty("privateEndpointIp")
+        private String privateEndpointIp;
+
+        public Builder privateEndpointIp(String privateEndpointIp) {
+            this.privateEndpointIp = privateEndpointIp;
+            this.__explicitlySet__.add("privateEndpointIp");
+            return this;
+        }
+
+        @com.fasterxml.jackson.annotation.JsonProperty("nsgIds")
+        private java.util.List<String> nsgIds;
+
+        public Builder nsgIds(java.util.List<String> nsgIds) {
+            this.nsgIds = nsgIds;
+            this.__explicitlySet__.add("nsgIds");
+            return this;
+        }
+
+        @com.fasterxml.jackson.annotation.JsonIgnore
+        private final java.util.Set<String> __explicitlySet__ = new java.util.HashSet<String>();
+
+        public PrivateEndpointDetails build() {
+            PrivateEndpointDetails __instance__ =
+                    new PrivateEndpointDetails(subnetId, privateEndpointIp, nsgIds);
+            __instance__.__explicitlySet__.addAll(__explicitlySet__);
+            return __instance__;
+        }
+
+        @com.fasterxml.jackson.annotation.JsonIgnore
+        public Builder copy(PrivateEndpointDetails o) {
+            Builder copiedBuilder =
+                    subnetId(o.getSubnetId())
+                            .privateEndpointIp(o.getPrivateEndpointIp())
+                            .nsgIds(o.getNsgIds());
+
+            copiedBuilder.__explicitlySet__.retainAll(o.__explicitlySet__);
+            return copiedBuilder;
+        }
+    }
+
+    /**
+     * Create a new builder.
+     */
+    public static Builder builder() {
+        return new Builder();
+    }
+
+    /**
+     * If specified, the stream pool will be private and only accessible from inside that subnet.
+     * Producing-to and consuming-from a stream inside a private stream pool can also only be done from inside the subnet.
+     * That value cannot be changed.
+     *
+     **/
+    @com.fasterxml.jackson.annotation.JsonProperty("subnetId")
+    String subnetId;
+
+    /**
+     * The optional private IP you want to be associated with your private stream pool.
+     * That parameter can only be specified when the subnetId parameter is set. It cannot be changed.
+     * The private IP needs to be part of the CIDR range of the specified subnetId or the creation will fail.
+     * If not specified a random IP inside the subnet will be chosen.
+     * After the stream pool is created, a custom FQDN, pointing to this private IP, is created.
+     * The FQDN is then used to access the service instead of the private IP.
+     *
+     **/
+    @com.fasterxml.jackson.annotation.JsonProperty("privateEndpointIp")
+    String privateEndpointIp;
+
+    /**
+     * The optional list of network security groups to be used with the private endpoint of the stream pool.
+     * That value cannot be changed.
+     *
+     **/
+    @com.fasterxml.jackson.annotation.JsonProperty("nsgIds")
+    java.util.List<String> nsgIds;
+
+    @com.fasterxml.jackson.annotation.JsonIgnore
+    private final java.util.Set<String> __explicitlySet__ = new java.util.HashSet<String>();
+}