+++ bmc-streaming/src/main/java/com/oracle/bmc/streaming/model/Stream.java
@@ -292,6 +292,8 @@ public class Stream {
 
     /**
      * The endpoint to use when creating the StreamClient to consume or publish messages in the stream.
+     * If the associated stream pool is private, the endpoint is also private and can only be accessed from inside the stream pool's associated subnet.
+     *
      **/
     @com.fasterxml.jackson.annotation.JsonProperty("messagesEndpoint")
     String messagesEndpoint;